Update README.md
RESPONSIVE WEB DESIGN: - Responsive design is a way to optimize the layout in a manner that all the significant information is presented in an optimal viewing way, with ease of reading and navigation with a minimum of resizing, panning and scrolling any kind of device or screen size. At WSNE consulting we offer excellent Responsive web design services making the website effective and easy to use on any device. This way there is no need for a different design and development phase when there is a new gadget on the market.
WSNE Consulting provides Responsive static website designing services that are unique and creative. Our Responsive static website designing strategy is very simple without complex programming. Basically these websites provide an overview of your business to your clients. Our professional team will assist you at each and every step. It will work for you to showcase your products and generate business.
WSNE Consulting Motive is to provide an attractive design to our valuable customers so that they can get a good response from the market. We always focus to maintain website loading speed so that your webpage could crawl fast within seconds. We have a wide range of professionals that can handle each and every situation; we consider every situation as an opportunity so that we can accomplish higher performance for our clients.
A Responsive static website contains web pages with fixed contents. Each and every page is coded with HTML and displays the relevant information to every visitor. It is the most basic website format and it doesn’t require any database design or web programming. Majorly its works well for small websites, but it is difficult to maintain with a large number of pages (Hundreds or Thousands). Static sites that contain a lot of pages are often designed using templates. You can update several pages at once and helps to provide a consistent layout through the site.
We have a web design approach aimed at crafting sites for screens of all kinds and sizes. The alluring, engaging and user-friendly designs have an impact on the users that makes them head to the website time and time.
WHY CHOOSE WSNE Consulting & WHY RESPONSIVE: -
•	TIME SAVING: - Responsive website design services offered by WSNE mainly cuts down on development time and cost. It allows you to optimize your business for all the devices with a single website.
•	LOADING TIME: - Loading time is another feature of responsive web design services offered by WSNE. You can reduce page load time with us as your browser doesn’t waste important seconds in redirecting to a mobile version.
•	SCREEN ADJUSTABLE: - As you browse a response site developed by WSNE with a mobile device there is no need to waste time with zooming, shrinking and pinching your screen. It automatically adjusts to your screen size and you get a quality user experience.
•	CUSTOMER SATISFACTION: - With WSNE, you get a website that looks clean, modern, sleek, and easy to ready. It creates customer satisfaction and builds trust and trust builds revenue.